requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
async-timeout==5.0.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
dateparser==1.2.2
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
pycryptodome==3.23.0
python-binance==1.0.36
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
regex==2026.1.15
requests==2.32.5
six==1.17.0
typing-extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
websockets==15.0.1
yarl==1.22.0
